Moving to replicate old new tab page
This makes a lot of changes that removes the extra crud that was
added to the new tab page.  Specifically, the extra bookmarks were
dropped (for now, stick with the bookmarks bar) along with the
top sites sidebar (this will come back).  The general theme has
been changed to mostly match the old new tab page.

The top sites bar is present, but hidden, as I have not yet figured
out how I want to do the separate areas.  I will try my best to
replicate the functionality to match the previous controls, but I
currently do not have the time and was only looking for the apps
to be displayed.

The new search, while incredibly useful for people with many apps,
was just another distraction for me, and has been dropped.  This
may make a comeback in the future, along with the ability to give
a custom order to apps on the page, similar to what could be done
in the original new tab page.

The settings and extra code that was added will eventually be
removed, but for now it pretty much ignores all of the settings
that are located in the modal.