Added left/right arrows for switching
This finally brings back a way to switch between the apps and top
sites section of the new tab page so they operate independently
of each other.  The keybindings and scrolling triggers have not
yet been implemented.

THe bookmarks section of the new tab page may return at a later
point, whether it is as a new section or implementing the other
bookmarks bar on the top.