Working on the new tab page footer
This starts the work on emulating the old new tab page's footer,
which included buttons to switch between tabs as well as links and
icons for things such as the Chrome web store.  The Chrome icon is
just the highly pixelated favicon until I can dig up the actual
logo from somewhere.