Include the favicon and site url in the top sites preview
Instead of just displaying the default application icon for the
top sites page, the icon has been swapped out with a grey box
that includes the favicon of the page as well as what can be shown
for the url.  This is the same as the original NTP, for what would
be displayed when a thumbnail could not be generated.

It is not possible to display the thumbnail for the website, as the
namspace is restricted to just the built-in NTP, and cannot be
accessed by extensions.